Move get_ipts_dir to the new location
Refs #336
