+++ tests/unit/test_utilities.py
@@ -1,7 +1,9 @@
 #!/usr/bin/python
 # Test utilities methods without GUI
 from __future__ import (absolute_import, division, print_function)  # python3 compatibility
+import os
 from pyrs.interface import gui_helper
+from pyrs.utilities.file_util import get_ipts_dir
 import pytest
 
 
@@ -20,6 +22,29 @@ def test_parse_integers():
     else:
         raise AssertionError('Shall be failed but get {0}'.format(int_list))
 
+
+@pytest.mark.skipif(not os.path.exists('/HFIR/HB2B/shared/'), reason='HFIR data archive is not mounted')
+def test_get_ipts():
+    """Test to get IPTS directory from run number
+
+    Returns
+    -------
+
+    """
+    # Test good
+    assert get_ipts_dir(1060) == '/HFIR/HB2B/IPTS-22731/', 'IPTS dir is not correct for run 1060'
+
+    # Test no such run
+    with pytest.raises(RuntimeError):
+        get_ipts_dir(112123260)
+
+    # Test exception
+    with pytest.raises(TypeError):
+        get_ipts_dir(1.2)
+    with pytest.raises(ValueError):
+        get_ipts_dir('1.2')
+    with pytest.raises(ValueError):
+        get_ipts_dir('abc')
 
 if __name__ == '__main__':
     pytest.main()
